scanner错误，无法完成

+++ lab02/src/adventure/SpeciesListStage.java
@@ -105,6 +105,6 @@ public class SpeciesListStage implements AdventureStage {
                 copy.remove(o);
             }
         }
-        return similarObjects / listOne.size();
+        return similarObjects / (double)listOne.size();
     }
 }
